removed harmless gcc warnings (unsigned vs signed int)

+++ vcg/complex/trimesh/clean.h
@@ -245,11 +245,11 @@ private:
 																					tri::Index(m,(*fi).V(2)),  
 																																										&*fi));
 						}
-				assert (m.fn == fvec.size());
+        assert (size_t(m.fn) == fvec.size());
 				//for(int i=0;i<fvec.size();++i) qDebug("fvec[%i] = (%i %i %i)(%i)",i,fvec[i].v[0],fvec[i].v[1],fvec[i].v[2],tri::Index(m,fvec[i].fp));
 				std::sort(fvec.begin(),fvec.end());
 				int total=0;
-				for(int i=0;i<fvec.size()-1;++i) 
+        for(size_t i=0;i<fvec.size()-1;++i)
 				{
 					if(fvec[i]==fvec[i+1])
 					{
